third_party/WebKit/LayoutTests/editing/deleting/delete_all_contents_with_link_crash.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>^foo</div>',
            '<div>bar <a href="http://www.apple.com/">baz|</a></div>',
        '</div>',
    ].join(''),
    'delete',
    [
        '<div contenteditable>',
            '|<br>',
        '</div>',
    ].join('')),
    'Delete all contents to placeholder');
</script>
